refactor routes.jsx structure
